requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
pydantic==1.10.2
pytz==2022.4
sqlparse==0.4.3
typing_extensions==4.4.0
